Feat: configure the number of days to lookback for
Closes: #22
