+++ github_lib.py
@@ -1,5 +1,7 @@
+import datetime
 import logging
 from functools import wraps
+from markupsafe import escape
 
 from flask import g, redirect, session, url_for, abort, render_template
 from flask_dance.contrib.github import github as gh_auth
@@ -9,8 +11,8 @@ from github import Github, Repository
 from github import UnknownObjectException, RateLimitExceededException, GithubException
 
 from utils import cache
-from errors import SAML403
-from config import GITHUB_ACCESS_TOKEN
+from errors import SAML403Exception
+from config import GITHUB_ACCESS_TOKEN, SINCE, MAX_COMMITS
 from base_routes import app
 
 # configure logging
@@ -38,7 +40,7 @@ def get_repo_contents(repo: Repository, path: str, cache_key: str) -> list:
         if isinstance(e, GithubException) and "SAML enforcement" in e.data.get(
             "message"
         ):
-            raise SAML403(config_repo)
+            raise SAML403Exception(config_repo)
         log.error(e, e.__traceback__)
         abort(500, "Error while listing files and folders")
 
@@ -63,9 +65,72 @@ def get_repo(g, config_repo: dict, cache_key: str) -> Repository:
         if isinstance(e, GithubException) and "SAML enforcement" in e.data.get(
             "message"
         ):
-            raise SAML403(config_repo)
+            raise SAML403Exception(config_repo)
         log.error(e, e.__traceback__)
         abort(500, description="Error while listing commits")
+
+
+@cached(cache)
+def get_commits(
+    repo: Repository,
+    section_path: str,
+    since: int = SINCE,
+    shared_token: bool = True,
+    cache_key: str = None,
+) -> list:
+    """Get the list of commits for the given repo and folder path.
+
+    Args:
+        repo (Repository): GitHub repository
+        section_path (str): path to the folder to monitor
+        since (int, optional): Number of days to look back. Defaults to SINCE
+        shared_token (bool, optional): Use the shared token or the user token. Defaults to True.
+        cache_key (str): token based key to use for the cache
+
+    Returns:
+        list: list of commits
+    """
+    log.debug(f"Looking for commits in {section_path}")
+    # get commit for the root of the repo requires no prefix slash
+    if section_path == "/":
+        section_path = ""
+    log.debug("Calculating the reference date")
+    ref_date = datetime.datetime.now() - datetime.timedelta(days=since)
+    try:
+        _commits = repo.get_commits(path=section_path, since=ref_date)
+    except RateLimitExceededException:
+        return abort(429, "Rate limit exceeded")
+    except Exception as e:
+        if isinstance(e, GithubException) and "SAML enforcement" in e.data.get(
+            "message"
+        ):
+            raise SAML403Exception(config_repo)
+        log.error(e, e.__traceback__)
+        return abort(500, "Error while listing commits")
+
+    log.debug(f"{_commits.totalCount} commits found in the last {since} days")
+    # Converting a limited list of commits
+    ret_commits = []
+    if _commits.totalCount > 0:
+        if _commits.totalCount > MAX_COMMITS and shared_token:
+            log.info("Using shared Github client: limiting commits to %s", MAX_COMMITS)
+            _commits = _commits[:MAX_COMMITS]
+        try:
+            for commit in _commits:
+                ret_commits.append(
+                    {
+                        "sha": escape(commit.sha[:7]),
+                        "author": escape(commit.commit.author.name),
+                        "commit": escape(commit.commit),
+                        "url": escape(commit.html_url),
+                        "message": escape(commit.commit.message),
+                        "date": commit.commit.author.date,
+                    }
+                )
+        except Exception as e:
+            log.error(e, e.__traceback__)
+            return abort(500, "Error while formatting commits")
+    return ret_commits
 
 
 def login_management(f):
