+++ utils.py
@@ -2,14 +2,21 @@
 """
 import datetime
 import logging
-from markupsafe import escape
 
 from github import Repository
 from flask import request, url_for, abort
 from feedgen.feed import FeedGenerator
 from cachetools import cached, TTLCache
 
-from config import *
+from config import (
+    CACHE_SIZE,
+    CACHE_TTL,
+    MAX_COMMITS,
+    APP_AUTHOR,
+    APP_AUTHOR_EMAIL,
+    APP_DESCRIPTION,
+    AZURE_DOCS_REPOS,
+)
 
 # Configure cache
 cache = TTLCache(maxsize=CACHE_SIZE, ttl=CACHE_TTL)
@@ -51,65 +58,6 @@ def get_repo_config(repo_owner: str = None, repo_name: str = None) -> dict:
             "icon": "",
         }
     return config_repo
-
-
-@cached(cache)
-def get_commits(
-    repo: Repository,
-    section_path: str,
-    since: int = SINCE,
-    shared_token: bool = True,
-    cache_key: str = None,
-) -> list:
-    """Get the list of commits for the given repo and folder path.
-
-    Args:
-        repo (Repository): GitHub repository
-        section_path (str): path to the folder to monitor
-        since (int, optional): Number of days to look back. Defaults to SINCE
-        shared_token (bool, optional): Use the shared token or the user token. Defaults to True.
-        cache_key (str): token based key to use for the cache
-
-    Returns:
-        list: list of commits
-    """
-    log.debug(f"Looking for commits in {section_path}")
-    # get commit for the root of the repo requires no prefix slash
-    if section_path == "/":
-        section_path = ""
-    log.debug("Calculating the reference date")
-    ref_date = datetime.datetime.now() - datetime.timedelta(days=since)
-    try:
-        _commits = repo.get_commits(path=section_path, since=ref_date)
-    except RateLimitExceededException:
-        return abort(429, "Rate limit exceeded")
-    except Exception as e:
-        log.error(e, e.__traceback__)
-        return abort(500, "Error while listing commits")
-
-    log.debug(f"{_commits.totalCount} commits found in the last {since} days")
-    # Converting a limited list of commits
-    ret_commits = []
-    if _commits.totalCount > 0:
-        if _commits.totalCount > MAX_COMMITS and shared_token:
-            log.info("Using shared Github client: limiting commits to %s", MAX_COMMITS)
-            _commits = _commits[:MAX_COMMITS]
-        try:
-            for commit in _commits:
-                ret_commits.append(
-                    {
-                        "sha": escape(commit.sha[:7]),
-                        "author": escape(commit.commit.author.name),
-                        "commit": escape(commit.commit),
-                        "url": escape(commit.html_url),
-                        "message": escape(commit.commit.message),
-                        "date": commit.commit.author.date,
-                    }
-                )
-        except Exception as e:
-            log.error(e, e.__traceback__)
-            return abort(500, "Error while formatting commits")
-    return ret_commits
 
 
 def get_feed(commits: list, folder: str, repo: dict) -> str:
